reset form.
The form will be reset after a successful password change.
