+++ src/Unosquare.PassCore.Web/ClientApp/types/Components.ts
@@ -1,0 +1,7 @@
+export interface IChangePasswordFormInitialModel {
+    CurrentPassword: string;
+    NewPassword: string;
+    NewPasswordVerify: string;
+    Recaptcha: string;
+    Username: string;
+}
